Fix for Settings app crash when toggle Mobile Data
Issue - The ExtTelephonymanager is declared as Static. It is created
and connected to serivce in OnCreate(). OnCreate() is called before
OnDestroy() is called when SIM Settings page is closed and opened
quickly, due to this the new OnCreate() will not be initializing
ExtTelephonymanager. When the old OnDestroy() is called, it makes the
ExtTelephonymanager disconnected and NULL.
Now the activity is left with NULL ExtTelephonymanager due to which
further queries using this is causing NPE. And the now OnDestroy()
called will cause NPE also.

FIX - Add change to query MSIM_CIWLAN support only once and reuse the
value instead of querying it every time. The OnDestroy will make
ExtTelephonymanager NULL only when there exists no instance of this
activity to prevent NPE when OnDestroy is called.

Change-Id: Ice94525c401f6040ab81326ef2b9cd5ee97d8fe2
CRs-Fixed: 3742178