Fix for pSIM/MEP profile switching on UI Settings
Currently ToggleSubscriptionDialogActivity will be finished
when the subId becomes inactive(i.e due to SIM removal case).

Since the explicit checking for SIM absent is not done, incase
where the PSIM + MEP switching is involved, and PSIM is supposed
to be turned ON, the MEP switch confirm dialog is getting dismissed
as soon as subscription changed event comes.Due to this, turning on
PSIM when both MEP profiles are in enabled state is failing.

Fix: Check for PSIM absent before finishing the ToggleSubscription
 -DialogActivity.

Change-Id: Ib3c7ab5b39a7ac4df33afd23e1f81bfd2f0c60cd
CRs-Fixed: 3801618