Fix NPE in Settings app.
satellite_attach_supported_bool carrier config is enabled for a carrier
but feature flag FEATURE_TELEPHONY_SATELLITE per device is not defined
so the crash is seen in settings app.

To fix the crash, null check is added to access Satellite service
which is enabled only if FEATURE_TELEPHONY_SATELLITE is defined.

Change-Id: Ia0bf26a28a7797f03e288352add068e0276c8bba
CRs-Fixed: 3841080